requirements.txt
ephem==3.7.5.1
requests==2.1.0
logging_tree==1.3
Markdown==2.2.1
# -e ~/src/django
django-debug-toolbar
